+++ src/main/java/lotto/utils/LottoWinningRank.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.utils;
 
 import java.util.Arrays;
 import java.util.List;
